Fix up routing
